requirements.txt
streamlit
sentence-transformers
torch
Pillow
pandas
requests
transformers
